refactor: support SSR
